+++ build-scripts/bundleDeps.js
@@ -10,12 +10,12 @@ const _ = require('lodash')
 const JS_DEPS = [
     {
         name: 'vue.min.js',
-        path: 'vue/dist/vue.min.js'
+        path: 'vue/dist/vue.runtime.global.prod.js',
     },
     {
         name: 'lodash.min.js',
-        path: 'lodash/lodash.min.js'
-    }
+        path: 'lodash/lodash.min.js',
+    },
 ]
 const WORKING_DIR = path.join(__dirname, '..')
 const NODE_MODULES_DIR = path.join(WORKING_DIR, 'node_modules')
